Move toolbox to UML package
