+++ gaphor/UML/toolbox.py
